moved tests to their own folder
